+++ beacon-chain/sync/initial-sync/blocks_fetcher_test.go
@@ -2,8 +2,6 @@ package initialsync
 
 import (
 	"context"
-	"fmt"
-	"math"
 	"sort"
 	"sync"
 	"testing"
@@ -12,14 +10,12 @@ import (
 	"github.com/kevinms/leakybucket-go"
 	core "github.com/libp2p/go-libp2p-core"
 	"github.com/libp2p/go-libp2p-core/network"
-	"github.com/libp2p/go-libp2p-core/peer"
 	eth "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
 	mock "github.com/prysmaticlabs/prysm/beacon-chain/blockchain/testing"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	dbtest "github.com/prysmaticlabs/prysm/beacon-chain/db/testing"
 	"github.com/prysmaticlabs/prysm/beacon-chain/flags"
 	p2pm "github.com/prysmaticlabs/prysm/beacon-chain/p2p"
-	"github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers/scorers"
 	p2pt "github.com/prysmaticlabs/prysm/beacon-chain/p2p/testing"
 	beaconsync "github.com/prysmaticlabs/prysm/beacon-chain/sync"
 	p2ppb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
@@ -28,7 +24,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
-	"github.com/prysmaticlabs/prysm/shared/timeutils"
 	"github.com/sirupsen/logrus"
 )
 
@@ -504,457 +499,6 @@ func TestBlocksFetcher_requestBeaconBlocksByRange(t *testing.T) {
 	assert.ErrorContains(t, "context canceled", err)
 }
 
-func TestBlocksFetcher_selectFailOverPeer(t *testing.T) {
-	type args struct {
-		excludedPID peer.ID
-		peers       []peer.ID
-	}
-	fetcher := newBlocksFetcher(context.Background(), &blocksFetcherConfig{})
-	tests := []struct {
-		name    string
-		args    args
-		want    peer.ID
-		wantErr error
-	}{
-		{
-			name: "No peers provided",
-			args: args{
-				excludedPID: "a",
-				peers:       []peer.ID{},
-			},
-			want:    "",
-			wantErr: errNoPeersAvailable,
-		},
-		{
-			name: "Single peer which needs to be excluded",
-			args: args{
-				excludedPID: "a",
-				peers: []peer.ID{
-					"a",
-				},
-			},
-			want:    "",
-			wantErr: errNoPeersAvailable,
-		},
-		{
-			name: "Single peer available",
-			args: args{
-				excludedPID: "a",
-				peers: []peer.ID{
-					"cde",
-				},
-			},
-			want:    "cde",
-			wantErr: nil,
-		},
-		{
-			name: "Two peers available, excluded first",
-			args: args{
-				excludedPID: "a",
-				peers: []peer.ID{
-					"a", "cde",
-				},
-			},
-			want:    "cde",
-			wantErr: nil,
-		},
-		{
-			name: "Two peers available, excluded second",
-			args: args{
-				excludedPID: "a",
-				peers: []peer.ID{
-					"cde", "a",
-				},
-			},
-			want:    "cde",
-			wantErr: nil,
-		},
-		{
-			name: "Multiple peers available",
-			args: args{
-				excludedPID: "a",
-				peers: []peer.ID{
-					"a", "cde", "cde", "cde",
-				},
-			},
-			want:    "cde",
-			wantErr: nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := fetcher.selectFailOverPeer(tt.args.excludedPID, tt.args.peers)
-			if tt.wantErr != nil {
-				assert.ErrorContains(t, tt.wantErr.Error(), err)
-			} else {
-				assert.Equal(t, tt.want, got)
-			}
-		})
-	}
-}
-
-func TestBlocksFetcher_nonSkippedSlotAfter(t *testing.T) {
-	peersGen := func(size int) []*peerData {
-		blocks := append(makeSequence(1, 64), makeSequence(500, 640)...)
-		blocks = append(blocks, makeSequence(51200, 51264)...)
-		blocks = append(blocks, 55000)
-		blocks = append(blocks, makeSequence(57000, 57256)...)
-		var peersData []*peerData
-		for i := 0; i < size; i++ {
-			peersData = append(peersData, &peerData{
-				blocks:         blocks,
-				finalizedEpoch: 1800,
-				headSlot:       57000,
-			})
-		}
-		return peersData
-	}
-	chainConfig := struct {
-		peers []*peerData
-	}{
-		peers: peersGen(5),
-	}
-
-	mc, p2p, _ := initializeTestServices(t, []uint64{}, chainConfig.peers)
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	fetcher := newBlocksFetcher(
-		ctx,
-		&blocksFetcherConfig{
-			finalizationFetcher: mc,
-			headFetcher:         mc,
-			p2p:                 p2p,
-		},
-	)
-	fetcher.rateLimiter = leakybucket.NewCollector(6400, 6400, false)
-	seekSlots := map[uint64]uint64{
-		0:     1,
-		10:    11,
-		31:    32,
-		32:    33,
-		63:    64,
-		64:    500,
-		160:   500,
-		352:   500,
-		480:   500,
-		512:   513,
-		639:   640,
-		640:   51200,
-		6640:  51200,
-		51200: 51201,
-	}
-	for seekSlot, expectedSlot := range seekSlots {
-		t.Run(fmt.Sprintf("range: %d (%d-%d)", expectedSlot-seekSlot, seekSlot, expectedSlot), func(t *testing.T) {
-			slot, err := fetcher.nonSkippedSlotAfter(ctx, seekSlot)
-			assert.NoError(t, err)
-			assert.Equal(t, expectedSlot, slot, "Unexpected slot")
-		})
-	}
-
-	t.Run("test isolated non-skipped slot", func(t *testing.T) {
-		seekSlot := uint64(51264)
-		expectedSlot := uint64(55000)
-		found := false
-		var i int
-		for i = 0; i < 100; i++ {
-			slot, err := fetcher.nonSkippedSlotAfter(ctx, seekSlot)
-			assert.NoError(t, err)
-			if slot == expectedSlot {
-				found = true
-				break
-			}
-		}
-		if !found {
-			t.Errorf("Isolated non-skipped slot not found in %d iterations: %v", i, expectedSlot)
-		} else {
-			t.Logf("Isolated non-skipped slot found in %d iterations", i)
-		}
-	})
-}
-
-func TestBlocksFetcher_filterPeers(t *testing.T) {
-	type weightedPeer struct {
-		peer.ID
-		usedCapacity int64
-	}
-	type args struct {
-		peers           []weightedPeer
-		peersPercentage float64
-	}
-	fetcher := newBlocksFetcher(context.Background(), &blocksFetcherConfig{})
-	tests := []struct {
-		name string
-		args args
-		want []peer.ID
-	}{
-		{
-			name: "no peers available",
-			args: args{
-				peers:           []weightedPeer{},
-				peersPercentage: 1.0,
-			},
-			want: []peer.ID{},
-		},
-		{
-			name: "single peer",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 10},
-				},
-				peersPercentage: 1.0,
-			},
-			want: []peer.ID{"a"},
-		},
-		{
-			name: "multiple peers same capacity",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 10},
-					{"b", 10},
-					{"c", 10},
-				},
-				peersPercentage: 1.0,
-			},
-			want: []peer.ID{"a", "b", "c"},
-		},
-		{
-			name: "multiple peers different capacity",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 20},
-					{"b", 15},
-					{"c", 10},
-					{"d", 90},
-					{"e", 20},
-				},
-				peersPercentage: 1.0,
-			},
-			want: []peer.ID{"c", "b", "a", "e", "d"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Non-leaking bucket, with initial capacity of 100.
-			fetcher.rateLimiter = leakybucket.NewCollector(0.000001, 100, false)
-			pids := make([]peer.ID, 0)
-			for _, pid := range tt.args.peers {
-				pids = append(pids, pid.ID)
-				fetcher.rateLimiter.Add(pid.ID.String(), pid.usedCapacity)
-			}
-			got, err := fetcher.filterPeers(pids, tt.args.peersPercentage)
-			require.NoError(t, err)
-			// Re-arrange peers with the same remaining capacity, deterministically .
-			// They are deliberately shuffled - so that on the same capacity any of
-			// such peers can be selected. That's why they are sorted here.
-			sort.SliceStable(got, func(i, j int) bool {
-				cap1 := fetcher.rateLimiter.Remaining(pids[i].String())
-				cap2 := fetcher.rateLimiter.Remaining(pids[j].String())
-				if cap1 == cap2 {
-					return pids[i].String() < pids[j].String()
-				}
-				return i < j
-			})
-			assert.DeepEqual(t, tt.want, got)
-		})
-	}
-}
-
-func TestBlocksFetcher_filterScoredPeers(t *testing.T) {
-	type weightedPeer struct {
-		peer.ID
-		usedCapacity int64
-	}
-	type args struct {
-		peers           []weightedPeer
-		peersPercentage float64
-		capacityWeight  float64
-	}
-
-	batchSize := uint64(flags.Get().BlockBatchLimit)
-	tests := []struct {
-		name   string
-		args   args
-		update func(s *scorers.BlockProviderScorer)
-		want   []peer.ID
-	}{
-		{
-			name: "no peers available",
-			args: args{
-				peers:           []weightedPeer{},
-				peersPercentage: 1.0,
-				capacityWeight:  0.2,
-			},
-			want: []peer.ID{},
-		},
-		{
-			name: "single peer",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 1200},
-				},
-				peersPercentage: 1.0,
-				capacityWeight:  0.2,
-			},
-			want: []peer.ID{"a"},
-		},
-		{
-			name: "multiple peers same capacity",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 2400},
-					{"b", 2400},
-					{"c", 2400},
-				},
-				peersPercentage: 1.0,
-				capacityWeight:  0.2,
-			},
-			want: []peer.ID{"a", "b", "c"},
-		},
-		{
-			name: "multiple peers capacity as tie-breaker",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 6000},
-					{"b", 3000},
-					{"c", 0},
-					{"d", 9000},
-					{"e", 6000},
-				},
-				peersPercentage: 1.0,
-				capacityWeight:  0.2,
-			},
-			update: func(s *scorers.BlockProviderScorer) {
-				s.IncrementProcessedBlocks("a", batchSize*2)
-				s.IncrementProcessedBlocks("b", batchSize*2)
-				s.IncrementProcessedBlocks("c", batchSize*2)
-				s.IncrementProcessedBlocks("d", batchSize*2)
-				s.IncrementProcessedBlocks("e", batchSize*2)
-			},
-			want: []peer.ID{"c", "b", "a", "e", "d"},
-		},
-		{
-			name: "multiple peers same capacity different scores",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 9000},
-					{"b", 9000},
-					{"c", 9000},
-					{"d", 9000},
-					{"e", 9000},
-				},
-				peersPercentage: 0.8,
-				capacityWeight:  0.2,
-			},
-			update: func(s *scorers.BlockProviderScorer) {
-				s.IncrementProcessedBlocks("e", s.Params().ProcessedBlocksCap)
-				s.IncrementProcessedBlocks("b", s.Params().ProcessedBlocksCap/2)
-				s.IncrementProcessedBlocks("c", s.Params().ProcessedBlocksCap/4)
-				s.IncrementProcessedBlocks("a", s.Params().ProcessedBlocksCap/8)
-				s.IncrementProcessedBlocks("d", 0)
-			},
-			want: []peer.ID{"e", "b", "c", "a"},
-		},
-		{
-			name: "multiple peers different capacities and scores",
-			args: args{
-				peers: []weightedPeer{
-					{"a", 6500},
-					{"b", 2500},
-					{"c", 1000},
-					{"d", 9000},
-					{"e", 6500},
-				},
-				peersPercentage: 0.8,
-				capacityWeight:  0.2,
-			},
-			update: func(s *scorers.BlockProviderScorer) {
-				// Make sure that score takes priority over capacity.
-				s.IncrementProcessedBlocks("c", batchSize*5)
-				s.IncrementProcessedBlocks("b", batchSize*15)
-				// Break tie using capacity as a tie-breaker (a and ghi have the same score).
-				s.IncrementProcessedBlocks("a", batchSize*3)
-				s.IncrementProcessedBlocks("e", batchSize*3)
-				// Exclude peer (peers percentage is 80%).
-				s.IncrementProcessedBlocks("d", batchSize)
-			},
-			want: []peer.ID{"b", "c", "a", "e"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mc, p2p, _ := initializeTestServices(t, []uint64{}, []*peerData{})
-			fetcher := newBlocksFetcher(context.Background(), &blocksFetcherConfig{
-				finalizationFetcher:      mc,
-				headFetcher:              mc,
-				p2p:                      p2p,
-				peerFilterCapacityWeight: tt.args.capacityWeight,
-			})
-			// Non-leaking bucket, with initial capacity of 10000.
-			fetcher.rateLimiter = leakybucket.NewCollector(0.000001, 10000, false)
-			peerIDs := make([]peer.ID, 0)
-			for _, pid := range tt.args.peers {
-				peerIDs = append(peerIDs, pid.ID)
-				fetcher.rateLimiter.Add(pid.ID.String(), pid.usedCapacity)
-			}
-			if tt.update != nil {
-				tt.update(fetcher.p2p.Peers().Scorers().BlockProviderScorer())
-			}
-			// Since peer selection is probabilistic (weighted, with high scorers having higher
-			// chance of being selected), we need multiple rounds of filtering to test the order:
-			// over multiple attempts, top scorers should be picked on high positions more often.
-			peerStats := make(map[peer.ID]int, len(tt.want))
-			var filteredPIDs []peer.ID
-			var err error
-			for i := 0; i < 1000; i++ {
-				filteredPIDs, err = fetcher.filterScoredPeers(context.Background(), peerIDs, tt.args.peersPercentage)
-				if len(filteredPIDs) <= 1 {
-					break
-				}
-				require.NoError(t, err)
-				for j, pid := range filteredPIDs {
-					// The higher peer in the list, the more "points" will it get.
-					peerStats[pid] += len(tt.want) - j
-				}
-			}
-
-			// If percentage of peers was requested, rebuild combined filtered peers list.
-			if len(filteredPIDs) != len(peerStats) && len(peerStats) > 0 {
-				filteredPIDs = []peer.ID{}
-				for pid := range peerStats {
-					filteredPIDs = append(filteredPIDs, pid)
-				}
-			}
-
-			// Sort by frequency of appearance in high positions on filtering.
-			sort.Slice(filteredPIDs, func(i, j int) bool {
-				return peerStats[filteredPIDs[i]] > peerStats[filteredPIDs[j]]
-			})
-			if tt.args.peersPercentage < 1.0 {
-				limit := uint64(math.Round(float64(len(filteredPIDs)) * tt.args.peersPercentage))
-				filteredPIDs = filteredPIDs[:limit]
-			}
-
-			// Re-arrange peers with the same remaining capacity, deterministically .
-			// They are deliberately shuffled - so that on the same capacity any of
-			// such peers can be selected. That's why they are sorted here.
-			sort.SliceStable(filteredPIDs, func(i, j int) bool {
-				score1 := fetcher.p2p.Peers().Scorers().BlockProviderScorer().Score(filteredPIDs[i])
-				score2 := fetcher.p2p.Peers().Scorers().BlockProviderScorer().Score(filteredPIDs[j])
-				if score1 == score2 {
-					cap1 := fetcher.rateLimiter.Remaining(filteredPIDs[i].String())
-					cap2 := fetcher.rateLimiter.Remaining(filteredPIDs[j].String())
-					if cap1 == cap2 {
-						return filteredPIDs[i].String() < filteredPIDs[j].String()
-					}
-				}
-				return i < j
-			})
-			assert.DeepEqual(t, tt.want, filteredPIDs)
-		})
-	}
-}
-
 func TestBlocksFetcher_RequestBlocksRateLimitingLocks(t *testing.T) {
 	p1 := p2pt.NewTestP2P(t)
 	p2 := p2pt.NewTestP2P(t)
@@ -1018,138 +562,6 @@ func TestBlocksFetcher_RequestBlocksRateLimitingLocks(t *testing.T) {
 		t.Error("p3 takes too long to respond: lock contention")
 	case <-ch:
 		// p3 responded w/o waiting for rate limiter's lock (on which p2 spins).
-	}
-}
-
-func TestBlocksFetcher_removeStalePeerLocks(t *testing.T) {
-	type peerData struct {
-		peerID   peer.ID
-		accessed time.Time
-	}
-	tests := []struct {
-		name     string
-		age      time.Duration
-		peersIn  []peerData
-		peersOut []peerData
-	}{
-		{
-			name:     "empty map",
-			age:      peerLockMaxAge,
-			peersIn:  []peerData{},
-			peersOut: []peerData{},
-		},
-		{
-			name: "no stale peer locks",
-			age:  peerLockMaxAge,
-			peersIn: []peerData{
-				{
-					peerID:   "a",
-					accessed: timeutils.Now(),
-				},
-				{
-					peerID:   "b",
-					accessed: timeutils.Now(),
-				},
-				{
-					peerID:   "c",
-					accessed: timeutils.Now(),
-				},
-			},
-			peersOut: []peerData{
-				{
-					peerID:   "a",
-					accessed: timeutils.Now(),
-				},
-				{
-					peerID:   "b",
-					accessed: timeutils.Now(),
-				},
-				{
-					peerID:   "c",
-					accessed: timeutils.Now(),
-				},
-			},
-		},
-		{
-			name: "one stale peer lock",
-			age:  peerLockMaxAge,
-			peersIn: []peerData{
-				{
-					peerID:   "a",
-					accessed: timeutils.Now(),
-				},
-				{
-					peerID:   "b",
-					accessed: timeutils.Now().Add(-peerLockMaxAge),
-				},
-				{
-					peerID:   "c",
-					accessed: timeutils.Now(),
-				},
-			},
-			peersOut: []peerData{
-				{
-					peerID:   "a",
-					accessed: timeutils.Now(),
-				},
-				{
-					peerID:   "c",
-					accessed: timeutils.Now(),
-				},
-			},
-		},
-		{
-			name: "all peer locks are stale",
-			age:  peerLockMaxAge,
-			peersIn: []peerData{
-				{
-					peerID:   "a",
-					accessed: timeutils.Now().Add(-peerLockMaxAge),
-				},
-				{
-					peerID:   "b",
-					accessed: timeutils.Now().Add(-peerLockMaxAge),
-				},
-				{
-					peerID:   "c",
-					accessed: timeutils.Now().Add(-peerLockMaxAge),
-				},
-			},
-			peersOut: []peerData{},
-		},
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	fetcher := newBlocksFetcher(ctx, &blocksFetcherConfig{})
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fetcher.peerLocks = make(map[peer.ID]*peerLock, len(tt.peersIn))
-			for _, data := range tt.peersIn {
-				fetcher.peerLocks[data.peerID] = &peerLock{
-					Mutex:    sync.Mutex{},
-					accessed: data.accessed,
-				}
-			}
-
-			fetcher.removeStalePeerLocks(tt.age)
-
-			var peersOut1, peersOut2 []peer.ID
-			for _, data := range tt.peersOut {
-				peersOut1 = append(peersOut1, data.peerID)
-			}
-			for peerID := range fetcher.peerLocks {
-				peersOut2 = append(peersOut2, peerID)
-			}
-			sort.SliceStable(peersOut1, func(i, j int) bool {
-				return peersOut1[i].String() < peersOut1[j].String()
-			})
-			sort.SliceStable(peersOut2, func(i, j int) bool {
-				return peersOut2[i].String() < peersOut2[j].String()
-			})
-			assert.DeepEqual(t, peersOut1, peersOut2, "Unexpected peers map")
-		})
 	}
 }
